// utils/graphql/storefront.graphql.gen.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  Any: { input: any; output: any };
  CEP: { input: any; output: any };
  CountryCode: { input: any; output: any };
  DateTime: { input: any; output: any };
  Decimal: { input: any; output: any };
  EmailAddress: { input: any; output: any };
  Long: { input: any; output: any };
  Upload: { input: any; output: any };
  Uuid: { input: any; output: any };
};
// ...
export type BuyBox = {
  /** List of the possibles installment plans. */
  installmentPlans?: Maybe<Array<Maybe<InstallmentPlan>>>;
  /** Maximum price among sellers. */
  maximumPrice?: Maybe<Scalars["Decimal"]["output"]>;
  /** Minimum price among sellers. */
  minimumPrice?: Maybe<Scalars["Decimal"]["output"]>;
  /** Quantity of offers. */
  quantityOffers?: Maybe<Scalars["Int"]["output"]>;
  /** List of sellers. */
  sellers?: Maybe<Array<Maybe<Seller>>>;
};
// ...
export type InstallmentPlan = {
  /** The custom display name of this installment plan. */
  displayName?: Maybe<Scalars["String"]["output"]>;
  /** List of the installments. */
  installments?: Maybe<Array<Maybe<Installment>>>;
  /** The name of this installment plan. */
  name?: Maybe<Scalars["String"]["output"]>;
};
// ...
export type Operation =
  /** Performs AND operation between query terms. */
  | "AND"
  /** Performs OR operation between query terms. */
  | "OR";
// ...
export type ProductFilterInput = {
  /** The attribute name. */
  field: Scalars["String"]["input"];
  /** The set of values which the result filter item value must be included in. */
  values: Array<InputMaybe<Scalars["String"]["input"]>>;
};
// ...
export type ProductSortKeys =
  /** The applied discount to the product variant price. */
  | "DISCOUNT"
  /** The product name. */
  | "NAME"
  /** The product variant price. */
  | "PRICE"
  /** Sort in a random way. */
  | "RANDOM"
  /** The date the product was released. */
  | "RELEASE_DATE"
  /** The sales number on a period of time. */
  | "SALES"
  /** The quantity in stock of the product variant. */
  | "STOCK";
// ...
export type Seller = {
  /** Seller name */
  name?: Maybe<Scalars["String"]["output"]>;
};
// ...
export type SortDirection =
  /** The results will be sorted in an ascending order. */
  | "ASC"
  /** The results will be sorted in an descending order. */
  | "DESC";
// ...
export type ProductFragment = {
  mainVariant?: boolean | null;
  productName?: string | null;
  productId?: any | null;
  alias?: string | null;
  available?: boolean | null;
  averageRating?: number | null;
  condition?: string | null;
  createdAt?: any | null;
  ean?: string | null;
  id?: string | null;
  minimumOrderQuantity?: number | null;
  productVariantId?: any | null;
  parentId?: any | null;
  sku?: string | null;
  numberOfVotes?: number | null;
  stock?: any | null;
  variantName?: string | null;
  variantStock?: any | null;
  collection?: string | null;
  urlVideo?: string | null;
  attributes?: Array<{
    value?: string | null;
    name?: string | null;
    attributeId?: string | null;
    displayType?: string | null;
    id?: string | null;
  } | null> | null;
  productCategories?: Array<{
    id: number;
    name?: string | null;
    url?: string | null;
    hierarchy?: string | null;
    main: boolean;
    googleCategories?: string | null;
  } | null> | null;
  informations?: Array<{
    title?: string | null;
    value?: string | null;
    type?: string | null;
  } | null> | null;
  images?: Array<{
    url?: string | null;
    fileName?: string | null;
    print: boolean;
  } | null> | null;
  prices?: {
    discountPercentage: any;
    discounted: boolean;
    listPrice?: any | null;
    multiplicationFactor: number;
    price: any;
    bestInstallment?: {
      discount: boolean;
      displayName?: string | null;
      fees: boolean;
      name?: string | null;
      number: number;
      value: any;
    } | null;
    installmentPlans?: Array<{
      displayName?: string | null;
      name?: string | null;
      installments?: Array<{
        discount: boolean;
        fees: boolean;
        number: number;
        value: any;
      } | null> | null;
    } | null> | null;
    priceTables?: Array<{
      discountPercentage: any;
      id: any;
      listPrice?: any | null;
      price: any;
    } | null> | null;
    wholesalePrices?: Array<{ price: any; quantity: number } | null> | null;
  } | null;
  productBrand?: {
    fullUrlLogo?: string | null;
    logoUrl?: string | null;
    name?: string | null;
    alias?: string | null;
  } | null;
  seller?: { name?: string | null } | null;
  similarProducts?: Array<{
    alias?: string | null;
    image?: string | null;
    imageUrl?: string | null;
    name?: string | null;
  } | null> | null;
  promotions?: Array<{
    content?: string | null;
    disclosureType?: string | null;
    id: any;
    fullStampUrl?: string | null;
    stamp?: string | null;
    title?: string | null;
  } | null> | null;
  buyBox?: BuyBox | null;
};
// ...
export type GetProductQueryVariables = Exact<{
  productId: Scalars["Long"]["input"];
  includeParentIdVariants?: InputMaybe<Scalars["Boolean"]["input"]>;
  includeAllVariants?: InputMaybe<Scalars["Boolean"]["input"]>;
  partnerAccessToken?: InputMaybe<Scalars["String"]["input"]>;
}>;
// ...
export type ProductPageQueryVariables = Exact<{
  filtersParent: ProductExplicitFiltersInput;
  firstParent: Scalars["Int"]["input"];
  sortDirectionParent: SortDirection;
  sortKeyParent?: InputMaybe<ProductSortKeys>;
  afterParent?: InputMaybe<Scalars["String"]["input"]>;
  filtersSecondary: ProductExplicitFiltersInput;
  firstSecondary: Scalars["Int"]["input"];
  sortDirectionSecondary: SortDirection;
  sortKeySecondary?: InputMaybe<ProductSortKeys>;
  afterSecondary?: InputMaybe<Scalars["String"]["input"]>;
  categoryId: Scalars["Long"]["input"];
  operationGenerics: Operation;
  queryGenerics?: InputMaybe<Scalars["String"]["input"]>;
  limitGenerics?: number;
  ignoreDisplayRulesGenerics?: boolean;
  maximumPriceGenerics?: InputMaybe<Scalars["Decimal"]["input"]>;
  filtersGenerics?: InputMaybe<
    Array<InputMaybe<ProductFilterInput>> | InputMaybe<ProductFilterInput>
  >;
  limitRecomendation: Scalars["Int"]["input"];
  url?: InputMaybe<Scalars["String"]["input"]>;
}>;
export interface ProductPageQuery {
  parentsProducts?: GetProductsQueryHome;
  productSecondary?: GetProductsQueryHome;
  productsCategories?: {
    nodes: Array<{
      name: string;
      hotsiteUrl: string;
      products: {
        edges: Array<{
          node: {
            productId: number;
            productName: string;
            alias: string;
            spotInformation: string;
            buyBox: {
              maximumPrice: number;
              minimumPrice: number;
            } | null;
          };
        }>;
      };
    }>;
  };
  productsGenerics?: {
    productsByOffset: {
      items: Array<{
        mainVariant: boolean;
        productName: string;
        productId: string;
        alias: string;
        id: string;
        productVariantId: string;
        parentId: string | null;
        sku: string;
        variantName: string | null;
        spotInformation: string | null;
        available: boolean;
        attributes: Array<{
          value: string;
          name: string;
        }>;
      }>;
    };
  };
  productRecomendations?: {
    productsByOffset?: {
      items?: Array<{
        mainVariant?: boolean | null;
        productName?: string | null;
        productId?: any | null;
        alias?: string | null;
        available?: boolean | null;
        averageRating?: number | null;
        condition?: string | null;
        createdAt?: any | null;
        ean?: string | null;
        id?: string | null;
        minimumOrderQuantity?: number | null;
        productVariantId?: any | null;
        parentId?: any | null;
        sku?: string | null;
        numberOfVotes?: number | null;
        stock?: any | null;
        variantName?: string | null;
        variantStock?: any | null;
        collection?: string | null;
        urlVideo?: string | null;
        attributes?: Array<{
          value?: string | null;
          name?: string | null;
        } | null> | null;
        productCategories?: Array<{
          id: number;
          name?: string | null;
          url?: string | null;
          hierarchy?: string | null;
          main: boolean;
          googleCategories?: string | null;
        } | null> | null;
        informations?: Array<{
          title?: string | null;
          value?: string | null;
          type?: string | null;
        } | null> | null;
        images?: Array<{
          url?: string | null;
          fileName?: string | null;
          print: boolean;
        } | null> | null;
        prices?: {
          discountPercentage: any;
          discounted: boolean;
          listPrice?: any | null;
          multiplicationFactor: number;
          price: any;
          bestInstallment?: {
            discount: boolean;
            displayName?: string | null;
            fees: boolean;
            name?: string | null;
            number: number;
            value: any;
          } | null;
          installmentPlans?: Array<{
            displayName?: string | null;
            name?: string | null;
            installments?: Array<{
              discount: boolean;
              fees: boolean;
              number: number;
              value: any;
            } | null> | null;
          } | null> | null;
          priceTables?: Array<{
            discountPercentage: any;
            id: any;
            listPrice?: any | null;
            price: any;
          } | null> | null;
          wholesalePrices?: Array<{
            price: any;
            quantity: number;
          } | null> | null;
        } | null;
        productBrand?: {
          fullUrlLogo?: string | null;
          logoUrl?: string | null;
          name?: string | null;
          alias?: string | null;
        } | null;
        seller?: { name?: string | null } | null;
        similarProducts?: Array<{
          alias?: string | null;
          image?: string | null;
          imageUrl?: string | null;
          name?: string | null;
        } | null> | null;
        promotions?: Array<{
          content?: string | null;
          disclosureType?: string | null;
          id: any;
          fullStampUrl?: string | null;
          stamp?: string | null;
          title?: string | null;
        } | null> | null;
        buyBox?: BuyBox | null;
      } | null> | null;
    } | null;
  };
}